+++ examples/config.py
@@ -1,0 +1,4 @@
+import os
+
+TARGET = os.getenv("TARGET", "https://www.osti.gov/elink2api/")
+TOKEN = os.getenv("TOKEN")